first clear

+++ Networking/MultiplayerPayloadFrame.cs
@@ -21,11 +21,13 @@ namespace CUCoreLib.Networking
                     putString.Invoke(null, new[] { writer, encoded, true });
                     return true;
                 }
+
                 var put = writer.GetType().GetMethod("Put", new[] { typeof(string) });
                 if (put == null) return false;
                 put.Invoke(writer, new object[] { encoded });
                 return true;
             }
+
             if (encoded.Length > MaxLength || putUShort == null || putUInt == null || putBytes == null)
                 throw new InvalidDataException("CUCoreLib payload exceeds the 16 MiB frame limit.");
             var bytes = Encoding.ASCII.GetBytes(encoded);
@@ -40,6 +42,7 @@ namespace CUCoreLib.Networking
                 Buffer.BlockCopy(bytes, offset, segment, 0, length);
                 putBytes.Invoke(writer, new object[] { segment });
             }
+
             return true;
         }
 
@@ -53,16 +56,18 @@ namespace CUCoreLib.Networking
                 getString.Invoke(null, args);
                 return args[1] as string;
             }
+
             var length = (ushort)getUShort.Invoke(reader, null);
             if (length != Sentinel)
             {
                 var segment = (ArraySegment<byte>)getSegment.Invoke(reader, new object[] { (int)length });
                 return Encoding.ASCII.GetString(segment.Array, segment.Offset, segment.Count);
             }
+
             var total = (uint)getUInt.Invoke(reader, null);
             var count = (ushort)getUShort.Invoke(reader, null);
             if (total <= SegmentLength || total > MaxLength || count != (total + SegmentLength - 1) / SegmentLength ||
-                availableBytes == null || (long)total + count * 2 > (int)availableBytes.GetValue(reader, null))
+                availableBytes == null || total + count * 2 > (int)availableBytes.GetValue(reader, null))
                 throw new InvalidDataException("Invalid segmented payload header or truncated frame.");
             var bytes = new byte[(int)total];
             var offset = 0;
@@ -70,10 +75,12 @@ namespace CUCoreLib.Networking
             {
                 var segment = (byte[])getBytes.Invoke(reader, null);
                 var lengthExpected = Math.Min(SegmentLength, bytes.Length - offset);
-                if (segment == null || segment.Length != lengthExpected) throw new InvalidDataException("Invalid segmented payload.");
+                if (segment == null || segment.Length != lengthExpected)
+                    throw new InvalidDataException("Invalid segmented payload.");
                 Buffer.BlockCopy(segment, 0, bytes, offset, segment.Length);
                 offset += segment.Length;
             }
+
             return Encoding.ASCII.GetString(bytes);
         }
     }
